image_resize_compress/requirements.txt
Pillow
typer